+++ src/alphanumeric.js
@@ -1,5 +1,0 @@
-import { numbers } from './numbers';
-import { lowercase } from './lowercase';
-import { uppercase } from './uppercase';
-
-export const alphanumeric = numbers + lowercase + uppercase;