+++ src/hexadecimal-lowercase.js
@@ -1,3 +1,0 @@
-import { numbers } from './numbers';
-
-export const hexadecimalLowercase = numbers + 'abcdef';