+++ src/hexadecimal-uppercase.js
@@ -1,3 +1,0 @@
-import { numbers } from './numbers';
-
-export const hexadecimalUppercase = numbers + 'ABCDEF';