+++ src/nanoid-dictionary.js
@@ -1,8 +1,0 @@
-export { alphanumeric } from './alphanumeric';
-export { hexadecimalLowercase } from './hexadecimal-lowercase';
-export { hexadecimalUppercase } from './hexadecimal-uppercase';
-export { lowercase } from './lowercase';
-export { nolookalikes } from './nolookalikes';
-export { nolookalikesSafe } from './nolookalikes-safe';
-export { numbers } from './numbers';
-export { uppercase } from './uppercase';