+++ src/nolookalikes.js
@@ -1,1 +1,0 @@
-export const nolookalikes = '346789ABCDEFGHJKLMNPQRTUVWXYabcdefghijkmnpqrtwxyz';