+++ src/numbers.js
@@ -1,1 +1,0 @@
-export const numbers = '0123456789';